requirements.txt
accelerate==0.25.0
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
datasets==2.16.0
dill==0.3.7
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.10.0
huggingface-hub==0.20.1
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.0
numpy==1.26.2
packaging==23.2
pandas==2.1.4
Pillow==9.3.0
psutil==5.9.7
pyarrow==14.0.2
pyarrow-hotfix==0.6
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.1
six==1.16.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchvision==0.16.2+cu121
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
xxhash==3.4.1
yarl==1.9.4
